refactor: remove provisioning manager
Greatly simplify the codebase by removing the provisioning
manager and moving the source into the main component
