--- components/network/CMakeLists.txt
@@ -1,1 +1,0 @@
-idf_component_register(SRC_DIRS "." INCLUDE_DIRS "." REQUIRES esp_event esp_http_client esp_wifi freertos log lwip nvs_flash wifi_provisioning)
