--- components/network/components.mk
@@ -1,4 +1,0 @@
-#
-# Component Makefile
-#
-COMPONENT_ADD_INCLUDEDIRS := .
