--- components/network/wifi.c
@@ -1,148 +1,0 @@
-#include "wifi.h"
-
-/* This tag will be prefixed to the log messages */
-static const char *TAG = "wifi";
-
-static const int WIFI_CONNECTED_EVENT = BIT0;
-static EventGroupHandle_t wifi_event_group;
-
-static void event_handler(void *arg, esp_event_base_t event_base, int32_t event_id, void *event_data) {
-    static int retries;
-    if (event_base == WIFI_PROV_EVENT) {
-        switch (event_id) {
-        case WIFI_PROV_START:
-            ESP_LOGI(TAG, "Provisioning started");
-            break;
-        case WIFI_PROV_CRED_RECV: {
-            wifi_sta_config_t *wifi_sta_cfg = (wifi_sta_config_t *) event_data;
-            ESP_LOGI(TAG, "Received WiFi credentials\n\tSSID: %s\n\tPassword: %s", (const char *) wifi_sta_cfg->ssid, (const char *) wifi_sta_cfg->password);
-            break;
-        }
-        case WIFI_PROV_CRED_FAIL: {
-            wifi_prov_sta_fail_reason_t *reason = (wifi_prov_sta_fail_reason_t *) event_data;
-            ESP_LOGE(TAG, "Provisioning failed!\n\tReason : %s\n\tPlease reset button and retry provisioning", (*reason == WIFI_PROV_STA_AUTH_ERROR) ? "WiFi station authentication failed" : "WiFi access point not found");
-            retries++;
-            if (retries >= 5) {
-                ESP_LOGI(TAG, "Failed to connect with provisioned AP, reseting provisioned credentials");
-                wifi_prov_mgr_reset_sm_state_on_failure();
-                retries = 0;
-            }
-            break;
-        }
-        case WIFI_PROV_CRED_SUCCESS:
-            ESP_LOGI(TAG, "Provisioning successful");
-            retries = 0;
-            break;
-        case WIFI_PROV_END:
-            wifi_prov_mgr_deinit(); /* De-initialize manager once provisioning is finished */
-            break;
-        default:
-            break;
-        }
-    } else if (event_base == WIFI_EVENT && event_id == WIFI_EVENT_STA_START) {
-        esp_wifi_connect();
-    } else if (event_base == IP_EVENT && event_id == IP_EVENT_STA_GOT_IP) {
-        ip_event_got_ip_t *event = (ip_event_got_ip_t *) event_data;
-        ESP_LOGI(TAG, "Connected with IP Address:" IPSTR, IP2STR(&event->ip_info.ip));
-        xEventGroupSetBits(wifi_event_group, WIFI_CONNECTED_EVENT); /* Signal main application to continue execution */
-    } else if (event_base == WIFI_EVENT && event_id == WIFI_EVENT_STA_DISCONNECTED) {
-        ESP_LOGI(TAG, "Disconnected. Connecting to the AP again...");
-        esp_wifi_connect();
-    }
-}
-
-static void setup_nvs(void) {
-    /* Initialize NVS partition */
-    esp_err_t ret = nvs_flash_init();
-    if (ret == ESP_ERR_NVS_NO_FREE_PAGES || ret == ESP_ERR_NVS_NEW_VERSION_FOUND) {
-        /* NVS partition was truncated
-         * and needs to be erased */
-        ESP_ERROR_CHECK(nvs_flash_erase());
-
-        /* Retry nvs_flash_init */
-        ESP_ERROR_CHECK(nvs_flash_init());
-    }
-}
-
-static void wifi_init_sta(void) {
-    /* Start Wi-Fi in station mode */
-    ESP_ERROR_CHECK(esp_wifi_set_mode(WIFI_MODE_STA));
-    ESP_ERROR_CHECK(esp_wifi_start());
-}
-
-static void provision_wifi(void) {
-    /* Configuration for the provisioning manager */
-    wifi_prov_mgr_config_t config = {.scheme = wifi_prov_scheme_softap, .scheme_event_handler = WIFI_PROV_EVENT_HANDLER_NONE};
-
-    /* Initialize provisioning manager with the
-     * configuration parameters set above */
-    ESP_ERROR_CHECK(wifi_prov_mgr_init(config));
-
-    bool provisioned = false;
-    /* WiFi credentials are saved in NVS and we could therefore be already provisioned. Let's check it out */
-    ESP_ERROR_CHECK(wifi_prov_mgr_is_provisioned(&provisioned));
-
-    /* If device is not yet provisioned start provisioning service */
-    if (!provisioned) {
-        ESP_LOGI(TAG, "Starting provisioning");
-
-        char *service_name;
-
-        if (!strcmp(CONFIG_WIFI_AP_NAME, "")) {
-            asprintf(&service_name, "%s", CONFIG_WIFI_AP_NAME);
-        } else {
-            uint8_t mac[6];
-            esp_read_mac(mac, ESP_MAC_WIFI_STA);
-            asprintf(&service_name, "ProgressBar-%X%X%X%X", mac[2], mac[3], mac[4], mac[5]);
-        }
-
-        wifi_prov_security_t security = WIFI_PROV_SECURITY_1;
-
-        /* Proof-of-Possesion */
-        const char *pop = CONFIG_WIFI_PROV_POP;
-
-        /* SoftAP password */
-        const char *service_key = CONFIG_WIFI_AP_PASS;
-
-        /* Start provisioning service */
-        ESP_ERROR_CHECK(wifi_prov_mgr_start_provisioning(security, pop, service_name, service_key));
-    } else {
-        ESP_LOGI(TAG, "Already provisioned, starting WiFi STA");
-
-        /* We don't need the manager as device is already provisioned,
-         * so let's release it's resources */
-        wifi_prov_mgr_deinit();
-
-        /* Start WiFi station */
-        wifi_init_sta();
-    }
-
-    /* Wait for WiFi connection */
-    xEventGroupWaitBits(wifi_event_group, WIFI_CONNECTED_EVENT, false, true, portMAX_DELAY);
-}
-
-void start_wifi(void) {
-    /* NVS for phy data and credentials */
-    setup_nvs();
-
-    /* Initialize TCP/IP */
-    ESP_ERROR_CHECK(esp_netif_init());
-
-    /* Initialize the event loop */
-    ESP_ERROR_CHECK(esp_event_loop_create_default());
-    wifi_event_group = xEventGroupCreate();
-
-    /* Register our event handler for WiFi, IP and Provisioning related events */
-    ESP_ERROR_CHECK(esp_event_handler_register(WIFI_PROV_EVENT, ESP_EVENT_ANY_ID, &event_handler, NULL));
-    ESP_ERROR_CHECK(esp_event_handler_register(WIFI_EVENT, ESP_EVENT_ANY_ID, &event_handler, NULL));
-    ESP_ERROR_CHECK(esp_event_handler_register(IP_EVENT, IP_EVENT_STA_GOT_IP, &event_handler, NULL));
-
-    /* Initialize WiFi including netif with default config */
-    esp_netif_create_default_wifi_sta();
-    esp_netif_create_default_wifi_ap();
-    wifi_init_config_t cfg = WIFI_INIT_CONFIG_DEFAULT();
-    ESP_ERROR_CHECK(esp_wifi_init(&cfg));
-
-    /* Now with the network stack running we can provision our connection */
-    provision_wifi();
-}
