+++ main/CMakeLists.txt
@@ -1,1 +1,1 @@
-idf_component_register(SRCS "app_main.c" INCLUDE_DIRS ".")
+idf_component_register(SRC_DIRS "." INCLUDE_DIRS ".")
