+++ main/app_main.h
@@ -2,6 +2,9 @@
 
 #include <esp_err.h>
 #include <esp_log.h>
+#include <esp_pm.h>
+
+#include <esp32c3/pm.h>
 
 #include "backend.h"
 #include "wifi.h"
