+++ main/wifi.h
@@ -3,17 +3,10 @@
 #include <stdio.h>
 #include <string.h>
 
-#include <freertos/FreeRTOS.h>
-#include <freertos/event_groups.h>
-#include <freertos/task.h>
-
 #include <esp_event.h>
 #include <esp_log.h>
 #include <esp_mac.h>
 #include <esp_wifi.h>
 #include <nvs_flash.h>
-
-#include <wifi_provisioning/manager.h>
-#include <wifi_provisioning/scheme_softap.h>
 
 void start_wifi(void);
